Add files via upload
The files here correspond with: the work I ddid on the Requirement specification on weeks 1,2,3. 

The Sprint1 Plan whcih we drafted collaboratively as a team on week4, based on the user stories I wrote .
 
The PSP AI Evidence which I made on day 4 and it represent my work on the AI during th sprint week, with the testing and comparison between the deafult AI and my pre-trained iterations

The Presentation witch the slides about the AI I made and delivered to our client on the final day of the sprint.